NixOS needs reboot logic + visual language
Add in nixos needs reboot logic. We've encoded a better/more robust
version of the script that actually flags the things I care about.

The pill now encodes the businesss logic of deciding severity in the
actual module and not at the aggregation level.i

The pill now encodes two levels of visual language.

First is determining icon. High value icons (in this case, media
playing) is the displayed icon. Otherwise, it will keep falling down
until it finds the right icon.

Second is color. Idle is properly represented with an "idle" color.
Second is info (say media is playing). And on for error/warning.